+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/PolarisLoadBalancerCompositeRuleTest.java
@@ -18,6 +18,7 @@
 
 package com.tencent.cloud.polaris.router;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
@@ -44,6 +45,10 @@ import com.tencent.cloud.polaris.loadbalancer.config.PolarisLoadBalancerProperti
 import com.tencent.cloud.polaris.router.config.properties.PolarisMetadataRouterProperties;
 import com.tencent.cloud.polaris.router.config.properties.PolarisNearByRouterProperties;
 import com.tencent.cloud.polaris.router.config.properties.PolarisRuleBasedRouterProperties;
+import com.tencent.cloud.polaris.router.interceptor.MetadataRouterRequestInterceptor;
+import com.tencent.cloud.polaris.router.interceptor.NearbyRouterRequestInterceptor;
+import com.tencent.cloud.polaris.router.interceptor.RuleBasedRouterRequestInterceptor;
+import com.tencent.cloud.polaris.router.spi.RouterRequestInterceptor;
 import com.tencent.polaris.api.pojo.DefaultInstance;
 import com.tencent.polaris.api.pojo.DefaultServiceInstances;
 import com.tencent.polaris.api.pojo.Instance;
@@ -87,23 +92,28 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	private PolarisRuleBasedRouterProperties polarisRuleBasedRouterProperties;
 	@Mock
 	private RouterAPI routerAPI;
+
 	private IClientConfig config;
 	private String testNamespace = "testNamespace";
 	private String testCallerService = "testCallerService";
 	private String testCalleeService = "testCalleeService";
 
+	private final List<RouterRequestInterceptor> requestInterceptors = new ArrayList<>();
+
 	@Before
 	public void before() {
 		config = new DefaultClientConfigImpl();
 		config.loadDefaultValues();
+		requestInterceptors.add(new MetadataRouterRequestInterceptor(polarisMetadataRouterProperties));
+		requestInterceptors.add(new NearbyRouterRequestInterceptor(polarisNearByRouterProperties));
+		requestInterceptors.add(new RuleBasedRouterRequestInterceptor(polarisRuleBasedRouterProperties));
 	}
 
 	@Test
 	public void testGetDefaultLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn("");
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule defaultRule = compositeRule.getRule();
 
@@ -114,8 +124,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void testRandomLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_RANDOM);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -126,8 +135,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void testWeightLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_WEIGHT);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -138,8 +146,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void testRetryLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_RETRY);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -150,8 +157,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void testWeightedResponseTimeLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_RESPONSE_TIME_WEIGHTED);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -162,8 +168,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void tesBestAvailableLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_BEST_AVAILABLE);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -174,8 +179,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void tesRoundRobinLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_ROUND_ROBIN);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -186,8 +190,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 	public void testAvailabilityFilteringLB() {
 		when(polarisLoadBalancerProperties.getStrategy()).thenReturn(PolarisLoadBalancerCompositeRule.STRATEGY_AVAILABILITY_FILTERING);
 		PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-				polarisLoadBalancerProperties, polarisNearByRouterProperties,
-				polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+				polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 		AbstractLoadBalancerRule lbRule = compositeRule.getRule();
 
@@ -206,13 +209,18 @@ public class PolarisLoadBalancerCompositeRuleTest {
 			setTransitiveMetadata();
 
 			PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-					polarisLoadBalancerProperties, polarisNearByRouterProperties,
-					polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+					polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 			ServiceInstances serviceInstances = assembleServiceInstances();
 			PolarisRouterContext routerContext = assembleRouterContext();
 
-			ProcessRoutersRequest request = compositeRule.buildProcessRoutersRequest(serviceInstances, routerContext);
+			Map<String, String> oldRouterLabels = routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS);
+			Map<String, String> newRouterLabels = new HashMap<>(oldRouterLabels);
+			newRouterLabels.put("system-metadata-router-keys", "k2");
+			routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, newRouterLabels);
+
+			ProcessRoutersRequest request = compositeRule.buildProcessRoutersBaseRequest(serviceInstances);
+			compositeRule.processRouterRequestInterceptors(request, routerContext);
 
 			Map<String, String> routerMetadata = request.getRouterMetadata(MetadataRouter.ROUTER_TYPE_METADATA);
 
@@ -236,13 +244,13 @@ public class PolarisLoadBalancerCompositeRuleTest {
 			setTransitiveMetadata();
 
 			PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-					polarisLoadBalancerProperties, polarisNearByRouterProperties,
-					polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+					polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 			ServiceInstances serviceInstances = assembleServiceInstances();
 			PolarisRouterContext routerContext = assembleRouterContext();
 
-			ProcessRoutersRequest request = compositeRule.buildProcessRoutersRequest(serviceInstances, routerContext);
+			ProcessRoutersRequest request = compositeRule.buildProcessRoutersBaseRequest(serviceInstances);
+			compositeRule.processRouterRequestInterceptors(request, routerContext);
 
 			Map<String, String> routerMetadata = request.getRouterMetadata(NearbyRouter.ROUTER_TYPE_NEAR_BY);
 
@@ -267,13 +275,13 @@ public class PolarisLoadBalancerCompositeRuleTest {
 			setTransitiveMetadata();
 
 			PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-					polarisLoadBalancerProperties, polarisNearByRouterProperties,
-					polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+					polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 			ServiceInstances serviceInstances = assembleServiceInstances();
 			PolarisRouterContext routerContext = assembleRouterContext();
 
-			ProcessRoutersRequest request = compositeRule.buildProcessRoutersRequest(serviceInstances, routerContext);
+			ProcessRoutersRequest request = compositeRule.buildProcessRoutersBaseRequest(serviceInstances);
+			compositeRule.processRouterRequestInterceptors(request, routerContext);
 
 			Map<String, String> routerMetadata = request.getRouterMetadata(RuleBasedRouter.ROUTER_TYPE_RULE_BASED);
 
@@ -298,8 +306,7 @@ public class PolarisLoadBalancerCompositeRuleTest {
 			setTransitiveMetadata();
 
 			PolarisLoadBalancerCompositeRule compositeRule = new PolarisLoadBalancerCompositeRule(routerAPI,
-					polarisLoadBalancerProperties, polarisNearByRouterProperties,
-					polarisMetadataRouterProperties, polarisRuleBasedRouterProperties, config);
+					polarisLoadBalancerProperties, config, requestInterceptors, null);
 
 			ProcessRoutersResponse assembleResponse = assembleProcessRoutersResponse();
 			when(routerAPI.processRouters(any())).thenReturn(assembleResponse);
@@ -339,8 +346,8 @@ public class PolarisLoadBalancerCompositeRuleTest {
 		Map<String, String> routerLabels = new HashMap<>();
 		routerLabels.put("k2", "v2");
 		routerLabels.put("k3", "v3");
-		routerContext.setLabels(PolarisRouterContext.TRANSITIVE_LABELS, transitiveLabels);
-		routerContext.setLabels(PolarisRouterContext.RULE_ROUTER_LABELS, routerLabels);
+		routerContext.putLabels(PolarisRouterContext.TRANSITIVE_LABELS, transitiveLabels);
+		routerContext.putLabels(PolarisRouterContext.ROUTER_LABELS, routerLabels);
 		return routerContext;
 	}