+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - [Bugfix: optimize ratelimit actuator](https://github.com/Tencent/spring-cloud-tencent/pull/413)
 - [Feature: add rate limit filter debug log](https://github.com/Tencent/spring-cloud-tencent/pull/417)
 - [Feature: Optimized configuration update](https://github.com/Tencent/spring-cloud-tencent/pull/423)
+- [Feature: add feature-env plugin & add spring cloud gateway staining plugin](https://github.com/Tencent/spring-cloud-tencent/pull/428)